Add RecordedBuild; modify Find*() to return it.
This new struct represents a build that ran. It associates the result of
build (success/failure) and an end time with a BuildId.

The Find*() functions return this now; since it embeds a *BuildId,
callers of those functions don't change much.

This sets us up for rendering a status HTML page, and potentially for
offering more info through the print/list commands.